+++ bindings/vue/test/e2e-webdriverio/dummy/index.spec.js
@@ -1,9 +1,0 @@
-describe('dummy', function() {
-  beforeEach(function() {
-    browser.url('/bindings/vue/test/e2e-webdriverio/dummy/index.html');
-  });
-
-  it('passes this test', function() {
-    expect(true).toBeTruthy();
-  });
-});